Allow multiple DBComponents
The existing implementation precluded multiple DBComponents in the
same view. I've changed `takeLatest` to `eachInline` to make sure
there are no issues with overlapping loads (we can add that in later),
and I've also removed the call that clears the local cache; this
should be done manually by the developer somewhere.