Adding basic pagination support
Pagination is now supported for next, previous, first, and last, only.
This isn't perfect, it would be better to have a range of pages
either side of the current, but it is a little harder to do generally.

This is a basic fix for pagination and only works for primary
queries, i.e. not many-to-many fields. In order to support those we'll
need to refactor a large chunk of the code.

Fixes #5